[ittapi] Fix ITTAPI download/build
Use `${ITTAPI_SOURCE_DIR}` rather than `${ITTAPI_SOURCE_DIR}/ittapi` for
the ITTAPI location.

Check out the `${ITTAPI_GIT_TAG}` only if ITTAPI is cloned because when
building from a tarball (as opposed to an online build),
`${ITTAPI_SOURCE_DIR}` is not a repository so we shouldn't try to check
out a branch.